Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int a;
-    int b;
+    float a;
+    float b;
     char ch;
-    scanf("%d %d %c",&a,&b,&ch);
+    scanf("%f %f %c",&a,&b,&ch);
     switch(ch){
         case '+':
             printf("%d",a+b);
